added in views to create and view each track. Starting on viewing and creating vehicles.

+++ db/migrate/20200712183929_create_tracks.rb
@@ -4,7 +4,7 @@ class CreateTracks < ActiveRecord::Migration[6.0]
       t.string :name
       t.string :location
       t.string :weather
-      t.integer :distance
+      t.decimal :distance
 
       t.timestamps
     end
